Separates queries from their execution
This allows users of the library to create CAPI queries without having
access to a configured API client - it breaks the dependency between
configuration, models and the execution of queries.
